# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MoodMargin API - meme coin perpetual trading demo platform with GenLayer risk intelligence
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type MarketVerdict = (typeof MarketVerdict)[keyof typeof MarketVerdict];

export const MarketVerdict = {
  WATCH: "WATCH",
  RESTRICT: "RESTRICT",
  AVOID: "AVOID",
  UNREVIEWED: "UNREVIEWED",
} as const;

export interface Market {
  id: string;
  symbol: string;
  name: string;
  tokenAddress?: string;
  chainName?: string;
  logoUrl?: string;
  coingeckoId?: string;
  dexPairAddress?: string;
  currentPrice?: number;
  priceChange24h?: number;
  volume24h?: number;
  liquidity?: number;
  marketCap?: number;
  verdict: MarketVerdict;
  maxLeverage?: number;
  tradingEnabled: boolean;
  riskScore?: number;
  createdAt?: string;
}

export interface PriceData {
  symbol: string;
  price: number;
  priceChange24h?: number;
  volume24h?: number;
  liquidity?: number;
  updatedAt?: string;
}

export interface MarketsSummary {
  totalMarkets: number;
  watchCount: number;
  restrictCount: number;
  avoidCount: number;
  unreviewedCount: number;
  totalVolume24h?: number;
  topGainer?: Market;
  topLoser?: Market;
}

export type RiskReviewRecommendation =
  (typeof RiskReviewRecommendation)[keyof typeof RiskReviewRecommendation];

export const RiskReviewRecommendation = {
  WATCH: "WATCH",
  RESTRICT: "RESTRICT",
  AVOID: "AVOID",
} as const;

export interface RiskReview {
  tokenAddress: string;
  chainName: string;
  tokenSymbol: string;
  reviewTimestamp?: number;
  riskScore: number;
  topHolderBps?: number;
  top10Bps?: number;
  ownershipStatus?: string;
  liquidityStatus?: string;
  deployerRiskNote?: string;
  recommendation: RiskReviewRecommendation;
  explanation?: string;
  aiExplanation?: string;
}

export interface SubmitTokenBody {
  tokenAddress: string;
  chainName: string;
  walletAddress: string;
}

export interface SubmitTokenResponse {
  success: boolean;
  review?: RiskReview;
  message: string;
}

export type ListingRequestStatus =
  (typeof ListingRequestStatus)[keyof typeof ListingRequestStatus];

export const ListingRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface ListingRequest {
  id: string;
  tokenAddress: string;
  chainName: string;
  tokenSymbol?: string;
  tokenName?: string;
  submittedBy: string;
  status: ListingRequestStatus;
  verdict?: string;
  notes?: string;
  createdAt: string;
}

export interface CreateListingRequestBody {
  tokenAddress: string;
  chainName: string;
  tokenSymbol?: string;
  tokenName?: string;
  submittedBy: string;
  notes?: string;
}

export interface ExplainRiskBody {
  tokenSymbol: string;
  recommendation: string;
  riskScore: number;
  explanation?: string;
  topHolderBps?: number;
  top10Bps?: number;
  ownershipStatus?: string;
  liquidityStatus?: string;
}

export interface AIExplanation {
  tokenSymbol: string;
  explanation: string;
  verdict: string;
  keyRisks?: string[];
  beginner_summary?: string;
}

export type PositionDirection =
  (typeof PositionDirection)[keyof typeof PositionDirection];

export const PositionDirection = {
  long: "long",
  short: "short",
} as const;

export type PositionStatus =
  (typeof PositionStatus)[keyof typeof PositionStatus];

export const PositionStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface Position {
  id: string;
  walletAddress: string;
  marketSymbol: string;
  marketName?: string;
  direction: PositionDirection;
  leverage: number;
  collateral: number;
  size: number;
  entryPrice: number;
  currentPrice?: number;
  liquidationPrice?: number;
  unrealizedPnl?: number;
  realizedPnl?: number;
  status: PositionStatus;
  openedAt?: string;
  closedAt?: string;
}

export type OpenPositionBodyDirection =
  (typeof OpenPositionBodyDirection)[keyof typeof OpenPositionBodyDirection];

export const OpenPositionBodyDirection = {
  long: "long",
  short: "short",
} as const;

export interface OpenPositionBody {
  walletAddress: string;
  marketSymbol: string;
  direction: OpenPositionBodyDirection;
  leverage: number;
  collateral: number;
}

export interface PnlSummary {
  walletAddress: string;
  totalRealizedPnl: number;
  totalUnrealizedPnl: number;
  totalPnl: number;
  winRate?: number;
  totalTrades?: number;
  winningTrades?: number;
  losingTrades?: number;
  bestTrade?: number;
  worstTrade?: number;
}

export interface FaucetStatus {
  walletAddress: string;
  canClaim: boolean;
  nextClaimAt?: string;
  lastClaimedAt?: string;
  totalClaimed?: number;
  claimAmount: number;
}

export interface ClaimFaucetBody {
  walletAddress: string;
}

export interface FaucetClaimResult {
  success: boolean;
  amount: number;
  newBalance: number;
  nextClaimAt?: string;
  message?: string;
}

export interface WalletProfile {
  walletAddress: string;
  mmUsdBalance: number;
  totalRealizedPnl?: number;
  totalTrades?: number;
  rank?: number;
  openPositionsCount?: number;
  createdAt?: string;
  lastSeenAt?: string;
}

export interface RegisterWalletBody {
  walletAddress: string;
}

export interface LeaderboardEntry {
  rank: number;
  walletAddress: string;
  totalPnl: number;
  totalTrades: number;
  winRate?: number;
  mmUsdBalance?: number;
  bestTrade?: number;
}

export type ListMarketsParams = {
  verdict?: ListMarketsVerdict;
  includeAvoid?: boolean;
};

export type ListMarketsVerdict =
  (typeof ListMarketsVerdict)[keyof typeof ListMarketsVerdict];

export const ListMarketsVerdict = {
  WATCH: "WATCH",
  RESTRICT: "RESTRICT",
  AVOID: "AVOID",
  ALL: "ALL",
} as const;

export type ListPositionsParams = {
  walletAddress: string;
  status?: ListPositionsStatus;
};

export type ListPositionsStatus =
  (typeof ListPositionsStatus)[keyof typeof ListPositionsStatus];

export const ListPositionsStatus = {
  open: "open",
  closed: "closed",
  all: "all",
} as const;

export type GetTradeHistoryParams = {
  walletAddress: string;
  limit?: number;
};

export type GetLeaderboardParams = {
  limit?: number;
  period?: GetLeaderboardPeriod;
};

export type GetLeaderboardPeriod =
  (typeof GetLeaderboardPeriod)[keyof typeof GetLeaderboardPeriod];

export const GetLeaderboardPeriod = {
  all: "all",
  week: "week",
  day: "day",
} as const;