# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MoodMargin API - meme coin perpetual trading demo platform with GenLayer risk intelligence
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /faucet/status/{walletAddress}:
    get:
      operationId: getFaucetStatus
      tags: [faucet]
      summary: Get faucet claim status for a wallet
      parameters:
        - in: path
          name: walletAddress
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Faucet status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaucetStatus"
  /faucet/claim:
    post:
      operationId: claimFaucet
      tags: [faucet]
      summary: Claim 1000 MMUSD daily faucet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimFaucetBody"
      responses:
        "200":
          description: Claim successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaucetClaimResult"
        "429":
          description: Already claimed today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ──────────────────────────────────────────
  # WALLET / USER
  # ──────────────────────────────────────────
  /wallet/{walletAddress}:
    get:
      operationId: getWalletProfile
      tags: [wallet]
      summary: Get user profile and demo balance
      parameters:
        - in: path
          name: walletAddress
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wallet profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletProfile"
  /wallet/register:
    post:
      operationId: registerWallet
      tags: [wallet]
      summary: Register or get a wallet profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterWalletBody"
      responses:
        "200":
          description: Wallet registered or returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletProfile"

  # ──────────────────────────────────────────
  # LEADERBOARD
  # ──────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    FaucetStatus:
      type: object
      properties:
        walletAddress:
          type: string
        canClaim:
          type: boolean
        nextClaimAt:
          type: string
        lastClaimedAt:
          type: string
        totalClaimed:
          type: number
        claimAmount:
          type: number
      required: [walletAddress, canClaim, claimAmount]
    ClaimFaucetBody:
      type: object
      properties:
        walletAddress:
          type: string
      required: [walletAddress]
    FaucetClaimResult:
      type: object
      properties:
        success:
          type: boolean
        amount:
          type: number
        newBalance:
          type: number
        nextClaimAt:
          type: string
        message:
          type: string
      required: [success, amount, newBalance]
    WalletProfile:
      type: object
      properties:
        walletAddress:
          type: string
        mmUsdBalance:
          type: number
        totalRealizedPnl:
          type: number
        totalTrades:
          type: integer
        rank:
          type: integer
        openPositionsCount:
          type: integer
        createdAt:
          type: string
        lastSeenAt:
          type: string
      required: [walletAddress, mmUsdBalance]
    RegisterWalletBody:
      type: object
      properties:
        walletAddress:
          type: string
      required: [walletAddress]
...